day_has_a_name/requirements.txt
-i https://pypi.org/simple
asgiref==3.2.7
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
django==3.0.4
google-api-core[grpc]==1.16.0
google-auth==1.12.0
google-cloud-secret-manager==0.2.0
googleapis-common-protos[grpc]==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.27.2
idna==2.9
mysqlclient==1.4.6
protobuf==3.11.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
